feat(workspace): port layer module to dasclaw_workspace_cap (#674 slice 2/N)
## 背景 / 目标

ADR-152 §3 F3.4 + 议题 #674：把 `desktop-client/ironclaw/src/workspace/` 剩余 9
个模块按"由叶到根、从小到大"逐切片迁入 `crates/dasclaw_workspace_cap/`。
切片 1/N（chunker, 200 行）已由 PR #673 完成；本 PR 是切片 2/N — `layer.rs`
(158 行)，零依赖的叶模块，结构与 chunker 切片完全对齐。

## 改动范围

- `git mv desktop-client/ironclaw/src/workspace/layer.rs
  crates/dasclaw_workspace_cap/src/layer.rs`（verbatim，字节未改）
- `crates/dasclaw_workspace_cap/src/lib.rs`：新增 `pub mod layer;`
- `desktop-client/ironclaw/src/workspace/layer.rs`：退化为 8 行 `pub use
  dasclaw_workspace_cap::layer::{LayerSensitivity, MemoryLayer};` 薄壳
- 14 处外部调用方（`config/channels.rs`、`config/workspace.rs`、`workspace/mod.rs`、
  `channels/web/tests/multi_tenant.rs`、`tests/layered_memory.rs` 等）零修改

## 非目标（What's NOT in this PR）

- 不重构 `MemoryLayer` API（包括 `default_for_user` / `read_scopes` /
  `writable_scopes` / `find` / `private_layer` 签名）
- 不改 serde 派生与 `rename_all = "snake_case"` 行为
- 不动剩余 8 个 workspace 子模块（privacy / document / embedding_cache /
  embeddings / hygiene / repository / search / mod）—后续切片
- 不更改 Cargo.toml 依赖集（serde 已在 dasclaw_workspace_cap 里）

## 验证

```
cargo check -p dasclaw_workspace_cap --tests          # 4.43s  OK
cargo nextest run -p dasclaw_workspace_cap            # 61/61 pass, 20.8s
cargo check -p dasclaw --tests                        # 4m32s  OK
cargo fmt --all                                       # clean
python3.12 scripts/check_no_panics.py --base origin/xClaw   # OK
cargo clippy --no-deps -p dasclaw_workspace_cap --all-targets -- -D warnings  # clean
cargo clippy --no-deps -p dasclaw --all-targets -- -D warnings                # 5m23s clean
```

## Sources read

- `AGENTS.md` §"红线" / "复用优先于新建"
- `docs/plans/architecture-refactor/adr-152-agent-and-capability-fusion.md` §3 F3.4（workspace 模块迁入清单）
- `docs/plans/architecture-refactor/adr-129-cross-crate-verbatim-port.md` §1.3（verbatim 移植约束）
- Issue #674（切片顺序与红线）
- PR #673（chunker 切片参考结构）
- `desktop-client/ironclaw/src/workspace/layer.rs`（原文件）
- `crates/dasclaw_workspace_cap/src/lib.rs` / `Cargo.toml`

Closes #691
Refs #674
Refs ADR-152 §3 F3.4
