+++ crates/dasclaw_workspace_cap/src/lib.rs
@@ -46,6 +46,7 @@
 //! ```
 
 pub mod chunker;
+pub mod layer;
 pub mod policy;
 pub use policy::{
     NetworkAccess, SandboxPolicy, WritableRoot, default_read_only_subpaths_for_writable_root,
